Moved the function that configures external dependencies (std::io) into the project submodule; because its role is to setup external dependencies for a project.
