+++ src/project.rs
@@ -1,7 +1,11 @@
 use std::path::{Path, PathBuf};
 
 use crate::{
-    ast::Module, diagnostics::config::TracingConfig, io::get_files, lexer::tokens::Token, parser,
+    ast::{Module, Type},
+    diagnostics::config::TracingConfig,
+    io::get_files,
+    lexer::tokens::Token,
+    parser,
 };
 
 use braid_lang::result::{NResult, Result};
@@ -201,4 +205,109 @@ fn truncate_extension(path: &mut Vec<String>, ext: &str) {
         Some(l) if l.ends_with(ext) => l.truncate(l.len() - ext.len()),
         _ => (),
     }
+}
+
+pub fn configure_imported_functions() -> Vec<(crate::ast::Path, Vec<Type>, Type)> {
+    vec![
+        (
+            vec!["root", "std", "io", "write"].into(),
+            vec![Type::StringLiteral],
+            Type::Unit,
+        ),
+        (
+            vec!["root", "std", "io", "readi64"].into(),
+            vec![],
+            Type::I64,
+        ),
+        (
+            vec!["root", "std", "io", "writeu8"].into(),
+            vec![Type::U8],
+            Type::Unit,
+        ),
+        (
+            vec!["root", "std", "io", "writeu8ln"].into(),
+            vec![Type::U8],
+            Type::Unit,
+        ),
+        (
+            vec!["root", "std", "io", "writeu16"].into(),
+            vec![Type::U16],
+            Type::Unit,
+        ),
+        (
+            vec!["root", "std", "io", "writeu16ln"].into(),
+            vec![Type::U16],
+            Type::Unit,
+        ),
+        (
+            vec!["root", "std", "io", "writeu32"].into(),
+            vec![Type::U32],
+            Type::Unit,
+        ),
+        (
+            vec!["root", "std", "io", "writeu32ln"].into(),
+            vec![Type::U32],
+            Type::Unit,
+        ),
+        (
+            vec!["root", "std", "io", "writeu64"].into(),
+            vec![Type::U64],
+            Type::Unit,
+        ),
+        (
+            vec!["root", "std", "io", "writeu64ln"].into(),
+            vec![Type::U64],
+            Type::Unit,
+        ),
+        (
+            vec!["root", "std", "io", "writei64"].into(),
+            vec![Type::I64],
+            Type::Unit,
+        ),
+        (
+            vec!["root", "std", "io", "writei64ln"].into(),
+            vec![Type::I64],
+            Type::Unit,
+        ),
+        (
+            vec!["root", "std", "io", "writei32"].into(),
+            vec![Type::I32],
+            Type::Unit,
+        ),
+        (
+            vec!["root", "std", "io", "writei32ln"].into(),
+            vec![Type::I32],
+            Type::Unit,
+        ),
+        (
+            vec!["root", "std", "io", "writei16"].into(),
+            vec![Type::I16],
+            Type::Unit,
+        ),
+        (
+            vec!["root", "std", "io", "writei16ln"].into(),
+            vec![Type::I16],
+            Type::Unit,
+        ),
+        (
+            vec!["root", "std", "io", "writei8"].into(),
+            vec![Type::I8],
+            Type::Unit,
+        ),
+        (
+            vec!["root", "std", "io", "writei8ln"].into(),
+            vec![Type::I8],
+            Type::Unit,
+        ),
+        (
+            vec!["root", "std", "io", "writebool"].into(),
+            vec![Type::Bool],
+            Type::Unit,
+        ),
+        (
+            vec!["root", "std", "io", "writeboolln"].into(),
+            vec![Type::Bool],
+            Type::Unit,
+        ),
+    ]
 }
